security

+++ Spring-Security/src/main/java/com/security/Spring/Security/appUser/repository/AppUserRepository.java
@@ -1,10 +1,10 @@
 package com.security.Spring.Security.appUser.repository;
 
-import com.security.Spring.Security.appUser.AppUser;
+import com.security.Spring.Security.appUser.model.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AppUserRepository extends JpaRepository<AppUser, Long> {
+public interface AppUserRepository extends JpaRepository<AppUser, Integer> {
     Optional<AppUser> findByEmail(String email);
 }
